FormationTemplateImageCreatorを削除
